feat(order): add migration edit total price column

+++ Services/Ordering/Ordering.Core/Entities/Order.cs
@@ -3,7 +3,7 @@
     public class Order : EntityBase
     {
         public string? UserName { get; set; }
-        public decimal? TotalPrice { get; set; }
+        public double? TotalPrice { get; set; }
         public string? FirstName { get; set; }
         public string? LastName { get; set; }
         public string? AddressLine { get; set; }
